Add function to round decimal result to two places.
-also replaced all parseInt with parseFloat

+++ app.js
@@ -87,13 +87,13 @@ numberSelection.forEach((button) => {
 let operate = function operation (num1, num2, operator){
 
     if (operator == "add") {
-        memory.runningSum = addCalc(parseInt(memory.num1), parseInt(memory.num2, operator));
+        memory.runningSum = addCalc(parseFloat(memory.num1), parseFloat(memory.num2, operator));
         memory.num1 = memory.runningSum;
         memory.num2 = undefined;
         console.log(memory);
     }
     else if (operator === "multiply") {
-        memory.runningSum = multiplyCalc(parseInt(memory.num1), parseInt(memory.num2, operator));
+        memory.runningSum = multiplyCalc(parseFloat(memory.num1), parseFloat(memory.num2, operator));
         memory.num1 = memory.runningSum;
         memory.num2 = undefined;
         console.log(memory);
@@ -113,7 +113,7 @@ equalsSelection.addEventListener('click', function (){
     else if (memory.runningSum == undefined) {
         console.log('no running sum');
         console.log(memory);
-        memory.sum = operate(parseInt(memory.num1), parseInt(memory.num2), operator);
+        memory.sum = operate(parseFloat(memory.num1), parseFloat(memory.num2), operator);
         memory.num2 = undefined;
         console.log(memory);
         
@@ -121,7 +121,7 @@ equalsSelection.addEventListener('click', function (){
     else if (memory.runningSum !==undefined) {
         console.log("there is a running sum. add num 1 and 2")
         console.log(memory);
-        memory.sum = operate(parseInt(memory.num1), parseInt(memory.num2), operator);
+        memory.sum = operate(parseFloat(memory.num1), parseFloat(memory.num2), operator);
         console.log(memory);
         memory.sum = undefined;
         memory.num2 = undefined;
